Flatten data structures
This commit is to simply flatten the Feature data structure in FeatureFloodCount.
The flat data structure is easier to iterate over in derived classes and
will make work towards reducing memory easier when only one rank is working
on merging. This commit should go in separately since it is large but the
intention is that it doesn't change _any_ existing behavior.

refs #7519